+++ sdk/storage/azure-storage-blobs/test/blob_batch_client_test.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/storage/blobs/blob.hpp"
+#include "azure/storage/blobs.hpp"
 #include "azure/storage/blobs/blob_sas_builder.hpp"
 #include "test_base.hpp"