+++ sdk/storage/azure-storage-blobs/test/blob_container_client_test.hpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/storage/blobs/blob.hpp"
+#include "azure/storage/blobs.hpp"
 #include "test_base.hpp"
 
 namespace Azure { namespace Storage { namespace Test {