+++ sdk/storage/azure-storage-common/test/bearer_token_test.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 #include "azure/identity/client_secret_credential.hpp"
-#include "azure/storage/blobs/blob.hpp"
+#include "azure/storage/blobs.hpp"
 #include "test_base.hpp"
 
 namespace Azure { namespace Storage { namespace Test {