+++ sdk/storage/azure-storage-files-datalake/test/datalake_directory_client_test.hpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/storage/files/datalake/datalake.hpp"
+#include "azure/storage/files/datalake.hpp"
 #include "datalake_path_client_test.hpp"
 #include "test_base.hpp"